requirements.txt
fastapi>=0.95
uvicorn>=0.22
python-dotenv>=1.0
requests>=2.30
pandas>=2.0
nltk>=3.8
pydantic>=2.5
sphinx>=8.2
myst-parser>=4.0
sphinx-rtd-theme
pylint==3.3.7
